requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
yt-dlp>=2023.10.07
openai-whisper>=20230308
ffmpeg-python>=0.2.0
python-dotenv>=0.19.0
pydub>=0.25.1
websockets>=10.4
python-dotenv>=0.19.0
python-slugify>=5.0.2
termcolor>=2.4.0